+++ JoyboyCommunity/src/components/Menu/index.tsx
@@ -17,7 +17,7 @@ import Animated, {
 } from 'react-native-reanimated';
 
 import {useStyles} from '../../hooks';
-import {Text} from '../_Text';
+import {Text} from '../Text';
 import stylesheet from './styles';
 
 export type MenuProps = {
