+++ JoyboyCommunity/src/components/index.ts
@@ -1,15 +1,15 @@
-export {Avatar, AvatarImage} from './_Avatar';
-export {Button} from './_Button';
-export {Divider} from './_Divider';
-export {Header} from './_Header';
-export {IconButton} from './_IconButton';
-export {Input} from './_Input';
-export {Menu} from './_Menu';
-export {KeyboardAvoidingView} from './_Skeleton/KeyboardAvoidingView';
-export {LayoutContainer} from './_Skeleton/LayoutContainer';
-export {RootScreenContainer} from './_Skeleton/RootScreenContainer';
-export {RowContainer} from './_Skeleton/RowContainer';
-export {ScreenContainer} from './_Skeleton/ScreenContainer';
-export {ScrollableContainer} from './_Skeleton/ScrollableContainer';
-export {Text} from './_Text';
+export {Avatar, AvatarImage} from './Avatar';
+export {Button} from './Button';
+export {Divider} from './Divider';
+export {Header} from './Header';
+export {IconButton} from './IconButton';
+export {Input} from './Input';
+export {Menu} from './Menu';
+export {KeyboardAvoidingView} from './Skeleton/KeyboardAvoidingView';
+export {LayoutContainer} from './Skeleton/LayoutContainer';
+export {RootScreenContainer} from './Skeleton/RootScreenContainer';
+export {RowContainer} from './Skeleton/RowContainer';
+export {ScreenContainer} from './Skeleton/ScreenContainer';
+export {ScrollableContainer} from './Skeleton/ScrollableContainer';
+export {Text} from './Text';
 export {TextButton} from './TextButton';
